refactor: move SRS and recognition helpers to utils
